move weightedRandomGame to separate file
